+++ kotlin/app/src/main/java/com/gamesage/kotlin/data/local/user/UserDao.kt
@@ -25,5 +25,5 @@ interface UserDao {
     fun observeAll(): Flow<List<UserEntity>>
 
     @Query("SELECT * FROM users WHERE id = :id")
-    suspend fun readUserById(id: Int): UserEntity?
+    suspend fun readUserById(id: Long): UserEntity?
 }
